fix: use .css extension on react-loading-skeleton import (#2362)
react-loading-skeleton defines an exports field in its package.json which looks like this:

"exports": {
    ".": {
      "types": "./dist/index.d.ts",
      "require": "./dist/index.cjs",
      "import": "./dist/index.js"
    },
    "./dist/skeleton.css": "./dist/skeleton.css"
  },

The webpack documentation states that if an exports field is defined in package.json, it replaces the default module loading behavior and any other request for a module will result in a ModuleNotFound error.

Our call to load react-loading-skeleton in Paragon looks like this:

@import "~react-loading-skeleton/dist/skeleton";

That isn’t in the list of possible exports.  This is though:

@import "~react-loading-skeleton/dist/skeleton.css";

This commit updates our import to the latter, which fixes downstream issues in MFEs that can’t figure out how to import react-loading-skeleton